Refactor with Subtopology
