+++ FlightExamples/ExamplesDeployment/Top/instances.fpp
@@ -125,7 +125,4 @@ module ExamplesDeployment {
 
   instance fprimeRouter: Svc.FprimeRouter base id 0x4D00
 
-  instance openSslWrapper: ExternalLibraryIntegration.OpenSslWrapper base id 0x4E00
-  
-  instance imageProcessor: ExternalLibraryIntegration.ImageProcessor base id 0x4F00
 }
