--- FlightExamples/ExternalLibraryIntegration/OpenSslWrapper/test/ut/OpenSslWrapperTestMain.cpp
@@ -1,27 +1,0 @@
-// ======================================================================
-// \title  OpenSslWrapperTestMain.cpp
-// \author thomas-bc
-// \brief  cpp file for OpenSslWrapper component test main function
-// ======================================================================
-
-#include "OpenSslWrapperTester.hpp"
-
-TEST(Nominal, testAesKeyRegistration) {
-    ExternalLibraryIntegration::OpenSslWrapperTester tester;
-    tester.testAesKeyRegistration();
-}
-
-TEST(Nominal, testAesDecryption) {
-    ExternalLibraryIntegration::OpenSslWrapperTester tester;
-    tester.testAesDecryption();
-}
-
-TEST(Nominal, testAesEncryption) {
-    ExternalLibraryIntegration::OpenSslWrapperTester tester;
-    tester.testAesEncryption();
-}
-
-int main(int argc, char** argv) {
-    ::testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
