+++ FlightExamples/ExternalLibs/ImageProcessor/test/ut/resources/fprime-logo.png
