+++ FlightExamples/ExternalLibs/Subtopology/CMakeLists.txt
@@ -1,0 +1,18 @@
+set(SOURCE_FILES
+    "${CMAKE_CURRENT_LIST_DIR}/intentionally-empty.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/ExternalLibsSubtopology-withOpenCV.fpp"
+    
+)
+# if (EXAMPLES_USE_OPENCV)
+#     list(APPEND SOURCE_FILES "${CMAKE_CURRENT_LIST_DIR}/ExternalLibsSubtopology-withOpenCV.fpp")
+# else()
+#     list(APPEND SOURCE_FILES "${CMAKE_CURRENT_LIST_DIR}/ExternalLibsSubtopology-noOpenCV.fpp")
+# endif ()
+
+register_fprime_module()
+
+set_target_properties(
+    ${FPRIME_CURRENT_MODULE}
+    PROPERTIES
+    SOURCES "${CMAKE_CURRENT_LIST_DIR}/intentionally-empty.cpp"
+)
