+++ FlightExamples/ExternalLibs/Subtopology/intentionally-empty.cpp
@@ -1,0 +1,7 @@
+/*
+This file is meant to be intentionally empty. Please do NOT delete this file.
+
+This file is included in the CMakeLists build source files, so that this subtopology can build
+on its own. This is in response to a bug with naming in the autocoded files, and may not be
+necessary in the future.
+*/
