+++ FlightExamples/SubtopologyConfigs/CMakeLists.txt
@@ -7,6 +7,7 @@
 
 set(SOURCE_FILES
   "${CMAKE_CURRENT_LIST_DIR}/ManagerWorkerConfig.fpp"
+  "${CMAKE_CURRENT_LIST_DIR}/ExternalLibsConfig.fpp"
 )
 
 register_fprime_module()
