+++ FlightExamples/project.cmake
@@ -4,7 +4,7 @@
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ManagerWorker/Manager/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ManagerWorker/Worker/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ManagerWorker/Subtopology/")
-add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ExternalLibraryIntegration/")
+add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ExternalLibs/")
 
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/SubtopologyConfigs/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ExamplesDeployment/")
